+++ Source/Core/Visualization/Renderer/StochasticPointRenderer.cpp
@@ -70,6 +70,21 @@ void StochasticPointRenderer::setDepthOffset( const float factor, const float un
     static_cast<Engine&>( engine() ).setDepthOffset( factor, units );
 }
 
+void StochasticPointRenderer::setVertexShaderFile( const std::string& file )
+{
+    static_cast<Engine&>( engine() ).setVertexShaderFile( file );
+}
+
+void StochasticPointRenderer::setFragmentShaderFile( const std::string& file )
+{
+    static_cast<Engine&>( engine() ).setFragmentShaderFile( file );
+}
+
+void StochasticPointRenderer::setShaderFiles( const std::string& vert_file, const std::string& frag_file )
+{
+    static_cast<Engine&>( engine() ).setShaderFiles( vert_file, frag_file );
+}
+
 /*===========================================================================*/
 /**
  *  @brief  Sets an opacity value.