+++ Source/Core/Visualization/Renderer/StochasticTetrahedraRenderer.h
@@ -47,6 +47,10 @@ public:
     void setEdgeFactor( const float factor );
     const kvs::TransferFunction& transferFunction() const;
     float samplingStep() const;
+    void setVertexShaderFile( const std::string& file );
+    void setGeometryShaderFile( const std::string& file );
+    void setFragmentShaderFile( const std::string& file );
+    void setShaderFiles( const std::string& vert_file, const std::string& geom_file, const std::string& frag_file );
 };
 
 /*===========================================================================*/
@@ -183,6 +187,17 @@ public:
 
     float samplingStep() const { return m_render_pass.samplingStep(); }
     const kvs::TransferFunction& transferFunction() const { return m_transfer_function; }
+
+    const std::string& vertexShaderFile() const { return m_render_pass.vertexShaderFile(); }
+    const std::string& geometryShaderFile() const { return m_render_pass.geometryShaderFile(); }
+    const std::string& fragmentShaderFile() const { return m_render_pass.fragmentShaderFile(); }
+    void setVertexShaderFile( const std::string& file ) { m_render_pass.setVertexShaderFile( file ); }
+    void setGeometryShaderFile( const std::string& file ) { m_render_pass.setGeometryShaderFile( file ); }
+    void setFragmentShaderFile( const std::string& file ) { m_render_pass.setFragmentShaderFile( file ); }
+    void setShaderFiles( const std::string& vert, const std::string& geom, const std::string& frag )
+    {
+        m_render_pass.setShaderFiles( vert, geom, frag );
+    }
 
 private:
     void create_transfer_function_texture();